Update file validation in download/hycom.py to avoind uneccisary loop.

+++ download/hycom.py
@@ -1,7 +1,7 @@
 import xarray as xr
 import cftime
 import pandas as pd
-import os
+import os,sys
 from datetime import datetime, timedelta
 import subprocess
 import numpy as np
@@ -37,17 +37,17 @@ def check_variables(file_path, expected_vars):
     """
     try:
         with xr.open_dataset(file_path) as ds:
-            missing_vars = [var for var in expected_vars if var not in ds.variables]
-            if not missing_vars:
-                print(f"All expected variables are present in {file_path}.")
-                return True
-            else:
-                print(f"Missing variables in {file_path}: {missing_vars}")
+            if expected_vars not in ds.variables:
+                print(f"Variable {expected_vars} is not in {file_path}.")
                 return False
+            else:
+                print(f"Variable {expected_vars} is present in {file_path}.")
+                return True
+                
     except Exception as e:
-        print(f"Error checking variables in file {file_path}: {e}")
+        print(f"Error checking variable in file {file_path}: {e}")
         return False
-        
+
 def check_time_range(file_path, expected_start, expected_end):
     """
     Function to check the time range of the file.
@@ -70,51 +70,44 @@ def check_time_range(file_path, expected_start, expected_end):
         return False
 
 def check_data_quality(file_path, variables):
-    """
-    Function to check the quality of the data downloaded.
-    """
-    try:
-        with xr.open_dataset(file_path) as ds:
-            for var in variables:
-                if var in ds:
-                    data = ds[var].values
-                    if data.size == 0:
-                        print(f"Variable {var} has no data in {file_path}.")
-                        return False
-                    else:
-                        pass
-
-                    if np.ndim(data) == 4:
-                        for dt in range(data[:,0,0,0].size):
-                            if np.isnan(data[dt]).all():
-                                print(f"All values for variable {var} at timestep {dt} are NaN in {file_path}.")
-                                return False
-                            else:
-                                pass
-                    elif np.ndim(data) == 3:
-                        for dt in range(data[:,0,0].size):
-                            if np.isnan(data[dt]).all():
-                                print(f"All values for variable {var} at timestep {dt} are NaN in {file_path}.")
-                                return False
-                            else:
-                                pass
-
-                    else:
-                        pass
-                        
-                else:
-                    print(f"Variable {var} not found in {file_path}.")
+                    print(f"Variable {var} has no data in {file_path}.")
                     return False
-                    
+                else:
+                    pass
+
+                if np.ndim(data) == 4:
+                    for dt in range(data[:,0,0,0].size):
+                        if np.isnan(data[dt]).all():
+                            print(f"All values for variable {var} at timestep {dt} are NaN in {file_path}.")
+                            return False
+                        else:
+                            pass
+
+                elif np.ndim(data) == 3:
+                    for dt in range(data[:,0,0].size):
+                        if np.isnan(data[dt]).all():
+                            print(f"All values for variable {var} at timestep {dt} are NaN in {file_path}.")
+                            return False
+                        else:
+                            pass
+
+                else:
+                    print(f"Variable {var} at timestep {dt} has incorrect dimension shape in {file_path}.")
+                    return False
+
+            else:
+                print(f"Variable {var} not found in {file_path}.")
+                return False
+
         return True
-        
+
     except Exception as e:
         print(f"Error checking data quality in file {file_path}: {e}")
-        return False
+
 
 def validate_download(file_path, expected_vars, expected_start, expected_end):
     """
-    Function that calls the validation routines to check the intergrity of the files downloaded. 
+    Function that calls the validation routines to check the intergrity of the files downloaded.
     """
     if not is_file_valid(file_path):
         return False
@@ -129,35 +122,7 @@ def validate_download(file_path, expected_vars, expected_start, expected_end):
     print(f"File {file_path} passed all validation checks.")
     return True
 
-
 def update_var_list(var_list):
-    var_metadata = {
-        'salinity': {
-            "vars": ["salinity"],
-            "url": "http://tds.hycom.org/thredds/dodsC/FMRC_ESPC-D-V02_s3z/FMRC_ESPC-D-V02_s3z_best.ncd",
-        },
-        'water_temp': {
-            "vars": ["water_temp"],
-            "url": "http://tds.hycom.org/thredds/dodsC/FMRC_ESPC-D-V02_t3z/FMRC_ESPC-D-V02_t3z_best.ncd",
-        },
-        'surf_el': {
-            "vars": ["surf_el"],
-            "url": "http://tds.hycom.org/thredds/dodsC/FMRC_ESPC-D-V02_ssh/FMRC_ESPC-D-V02_ssh_best.ncd",
-        },
-        'water_u': {
-            "vars": ["water_u"],
-            "url": "http://tds.hycom.org/thredds/dodsC/FMRC_ESPC-D-V02_u3z/FMRC_ESPC-D-V02_u3z_best.ncd",
-        },
-        'water_v': {
-            "vars": ["water_v"],
-            "url": "http://tds.hycom.org/thredds/dodsC/FMRC_ESPC-D-V02_v3z/FMRC_ESPC-D-V02_v3z_best.ncd",
-        }
-    }
-
-    return {var: var_metadata[var] for var in var_list if var in var_metadata}
-
-
-def decode_time_units(time_var):
     try:
         units = time_var.units
         calendar = getattr(time_var, 'calendar', 'standard')
@@ -173,7 +138,7 @@ def decode_time_units(time_var):
 
 def download_var(var, metadata, domain, depths, save_dir, run_date, hdays, fdays):
     vars_to_drop = ['salinity_bottom', 'water_temp_bottom', 'water_u_bottom', 'water_v_bottom', 'tau', 'time_offset',
-                    'time_run', 'time1_offset', 'sst', 'sss', 'ssu', 'ssv', 'sic', 'sih', 'siu', 'siv', 'surtx', 
+                    'time_run', 'time1_offset', 'sst', 'sss', 'ssu', 'ssv', 'sic', 'sih', 'siu', 'siv', 'surtx',
                     'surty', 'steric_ssh']
     lon_range, lat_range, depth_range = slice(domain[0], domain[1]), slice(domain[2], domain[3]), slice(depths[0], depths[1])
 
@@ -181,7 +146,7 @@ def download_var(var, metadata, domain, depths, save_dir, run_date, hdays, fdays
     start_date = pd.Timestamp(run_date) - timedelta(days=hdays)
     end_date = pd.Timestamp(run_date) + timedelta(days=fdays)
     time_range = slice(start_date, end_date)
-    
+
     MAX_RETRIES, RETRY_WAIT = 3, 20
     variable=None
     for attempt in range(MAX_RETRIES):
@@ -189,59 +154,71 @@ def download_var(var, metadata, domain, depths, save_dir, run_date, hdays, fdays
         print(f'Attempt {attempt} out of {MAX_RETRIES} tries for {metadata["vars"][0]}.')
         try:
             print(f'Connecting to {metadata["url"]} to subset and download {metadata["vars"][0]}.')
-            ds = xr.open_dataset(metadata["url"], 
-                                 drop_variables=vars_to_drop, 
-                                 decode_times=False, 
-                                 engine="netcdf4").sel(lat=lat_range, 
+            ds = xr.open_dataset(metadata["url"],
+                                 drop_variables=vars_to_drop,
+                                 decode_times=False,
+                                 engine="netcdf4").sel(lat=lat_range,
                                                        lon=lon_range)
             if 'time' in ds:
                 ds['time'] = decode_time_units(ds['time'])
                 ds = ds.sel(time=time_range)
-            
+
             variable = ds[metadata["vars"][0]]
             if variable.ndim == 4:
                 variable = variable.sel(depth=depth_range)
-            
+
             variable = variable.resample(time='1D').mean()
             #variable = variable.resample(time='1D', offset='12h').mean()
             save_path = os.path.join(save_dir, f"hycom_{metadata['vars'][0]}.nc")
             variable.to_netcdf(save_path, 'w')
             print(f'File written to {save_path}')
             ds.close()
-            break
-            
+
+            if validate_download(save_path, metadata["vars"][0], start_date, end_date):
+                print('')
+                print(f"File {save_path} and validated successfully!")
+                print('')
+                break
+
+            else:
+                print('')
+                print(f"File {save_path} validation failed and retrying the download")
+                print('')
+
         except Exception as e:
             print(f"Error: {e}")
             if attempt < MAX_RETRIES - 1:
                 print(f"Retrying in {RETRY_WAIT} seconds...")
                 time.sleep(RETRY_WAIT)
             else:
-                print("Failed to download after multiple attempts.")
-                
+                print(f"Failed to download after the maximum number of attempts: {MAX_RETRIES}.")
+
         finally:
             # Explicitly delete large variables to free up memory
             del variable
 
 def download_vars_parallel(variables, domain, depths, run_date, hdays, fdays, workers, save_dir):
     var_metadata = update_var_list(variables)
-    
+
     with ThreadPoolExecutor(max_workers=workers) as executor:
         future_to_var = {
-            executor.submit(download_var, var, metadata, domain, depths, save_dir, run_date, hdays, fdays): var 
+            executor.submit(download_var, var, metadata, domain, depths, save_dir, run_date, hdays, fdays): var
             for var, metadata in var_metadata.items()
         }
-        
+
         for future in as_completed(future_to_var):
             var = future_to_var[future]
             try:
                 future.result()
+                return True
             except Exception as e:
                 print(f"Download failed for {var}: {e}")
+                return False
 
 def download_hycom(variables, domain, depths, run_date, hdays, fdays, save_dir, workers=None):
     """
-    Downloads the HYCOM analysis in daily outputs using xarrray opendap. 
-    This function does check the integrity of the file. If the file is corrupt the download retries - max retries is 3.  
+    Downloads the HYCOM analysis in daily outputs using xarrray opendap.
+    This function does check the integrity of the file. If the file is corrupt the download retries - max retries is 3.
 
     INPUTS:
     variables: List of variables to download (e.g. ['salinity', 'water_temp', 'surface_el', 'water_u', 'water_v'])
@@ -251,67 +228,30 @@ def download_hycom(variables, domain, depths, run_date, hdays, fdays, save_dir, 
     hdays    : Days to hindcast (e.g. hdays=5).
     fdays    : Days to forecast (e.g. fdays=5).
     save_dir : Directory to save the downloaded data (eg. save_dir='/path/and/directory/to/save/').
-    workers  : It is the number of variables to download in parallel. Default is None, in which cases it downloads all the variables in paralell. 
+    workers  : It is the number of variables to download in parallel. Default is None, in which cases it downloads all the variables in paralell.
                To note, the number of workers cannot exceed the number of variables.
 
     OUTPUT:
     NetCDF file containing the most recent HYCOM forcast run.
     """
-    
-    # We add an additional day to ensure that it exceeds the model run time. 
+
+    # We add an additional day to ensure that it exceeds the model run time.
     hdays, fdays = hdays+1, fdays+1
-    
+
     if workers is None:
         workers=len(variables)
     else:
         pass
-    
-    # We initialise a loop to retry the download if the file was corrupt. 
-    # The MAX_TRIES is set at 3 to avoid an infinate loop.
-    MAX_TRIES=3
-    for attempt in range(MAX_TRIES):
-        print( '--------------------------------------------')
-        print(f'    Attempt {attempt} out of {MAX_TRIES}    ')
-        print( '--------------------------------------------')
 
-        download_vars_parallel(variables, domain, depths, run_date, hdays, fdays, workers, save_dir)
-    
+    if download_vars_parallel(variables, domain, depths, run_date, hdays, fdays, workers, save_dir):
         ds = xr.open_mfdataset(os.path.join(save_dir, 'hycom_*.nc'))
-        
         outfile = os.path.abspath(os.path.join(save_dir, f"HYCOM_{run_date.strftime('%Y%m%d_%H')}.nc"))
-        
         if os.path.exists(outfile):
             os.remove(outfile)
-        
         ds.to_netcdf(outfile, 'w')
-        
         subprocess.call(["chmod", "-R", "775", outfile])
         print('')
-        print('Validating download...')
+        print('created: ', outfile)
         print('')
-        
-        start_date = pd.Timestamp(run_date) - timedelta(days=hdays)
-        end_date = pd.Timestamp(run_date) + timedelta(days=fdays)
-        
-        if validate_download(outfile, variables, start_date, end_date):
-            print('')
-            print("File downloaded and validated successfully!")
-            print('')
-            print('created: ', outfile)
-            print('')
-            break
-        else:
-            print('')
-            print("File validation failed and retrying the download")
-            print('')
-
-
-if __name__ == '__main__':
-    run_date = pd.to_datetime('2024-12-02 12:00:00')
-    hdays = 5
-    fdays = 5
-    variables = ['salinity','water_temp','surf_el','water_u','water_v']
-    domain = [23,25,-37,-36]
-    depths = [0,5000]
-    save_dir = '/home/g.rautenbach/Projects/somisana-croco/DATASETS_CROCOTOOLS/HYCOM/'
-    download_hycom(variables, domain, depths, run_date, hdays, fdays, save_dir, workers=None)
+    else:
+        print('HYCOM download failed.')
